Mailbox System Update
The way internal messages are sent as events has been updated. There is now 1 event for application messages. The message sent with it has an event ID in it to determine what sort of message it is. Along with that is a pointer to data. This data is free'd up once the message has been processed.